+++ tests/tests_api_v1_1/test_api_v1_1_auth.py
@@ -1,7 +1,8 @@
 # coding: utf-8
+import pendulum
 from rest_framework.test import APITestCase
 from rest_framework import status
-from django.contrib.auth import authenticate
+from django.contrib.auth import authenticate, get_user_model
 from django.test import Client
 from concrete_datastore.concrete.models import User, UserConfirmation
 from django.test import override_settings
@@ -13,14 +14,26 @@ class AuthTestCase(APITestCase):
         # Create a user
         self.user = User.objects.create_user(
             email='johndoe@netsach.org',
-            password='plop'
+            password='plop',
             # 'John',
             # 'Doe',
         )
-        self.user.save()
         self.confirmation = UserConfirmation.objects.create(user=self.user)
         self.confirmation.confirmed = True
         self.confirmation.save()
+
+        # Create a blocked user
+        self.user2 = User.objects.create_user(
+            email='janedoe@netsach.org', password='plop'
+        )
+        self.user2.save()
+        self.user2.level = 'blocked'
+
+        # Create an unconfirmed user
+        self.user3 = User.objects.create_user(
+            email='juliadoe@netsach.org', password='plop'
+        )
+        self.user3.save()
 
     def test_authenticate_user_cant_authenticate(self):
         self.user.is_staff = True
@@ -57,11 +70,7 @@ class AuthTestCase(APITestCase):
 
         # POST a valid user/password and get a session ID (200)
         resp = self.client.post(
-            url,
-            {
-                "email": 'johndoe@netsach.org',
-                "password": "plop",
-            },
+            url, {"email": 'johndoe@netsach.org', "password": "plop",},
         )
         self.assertEqual(
             resp.status_code, status.HTTP_200_OK, msg=resp.content
@@ -74,11 +83,7 @@ class AuthTestCase(APITestCase):
         self.assertIn('groups', resp.data)
 
         resp = self.client.post(
-            url,
-            {
-                "email": 'JOHNDOE@netsach.org',
-                "password": "plop",
-            },
+            url, {"email": 'JOHNDOE@netsach.org', "password": "plop",},
         )
 
         # print(resp.status_code)
@@ -87,11 +92,7 @@ class AuthTestCase(APITestCase):
         )
 
         resp = self.client.post(
-            url,
-            {
-                "email": 'johndoe@netsach.org',
-                "password": "plop",
-            },
+            url, {"email": 'johndoe@netsach.org', "password": "plop",},
         )
 
         # print(resp.status_code)
@@ -126,15 +127,72 @@ class AuthTestCase(APITestCase):
         # POST an invalid arguimentsand get an error (400)
         resp = self.client.post(url, {"email": 'johndoe@netsach.org'})
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['INVALID_DATA'])
 
         # POST an invalid email and get an error (400)
         resp = self.client.post(
             url, {"email": 'johndoeXXXX', "password": "wrong-password"}
         )
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['INVALID_DATA'])
 
-        # POST an invalid user/password and get an error (401)
+        # POST an unknown email and get an error (401)
+        resp = self.client.post(
+            url, {"email": 'johndoe2@netsach.org', "password": "some-password"}
+        )
+        self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['WRONG_AUTH_CREDENTIALS'])
+
+        # POST an invalid password and get an error (401)
         resp = self.client.post(
             url, {"email": 'johndoe@netsach.org', "password": "wrong-password"}
         )
         self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['WRONG_AUTH_CREDENTIALS'])
+
+        # POST as a blocked user
+        resp = self.client.post(
+            url, {"email": 'janedoe@netsach.org', "password": "plop"}
+        )
+        self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['USER_BLOCKED'])
+
+        # POST as a blocked user
+        resp = self.client.post(
+            url, {"email": 'janedoe@netsach.org', "password": "plop"}
+        )
+        self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['USER_BLOCKED'])
+
+    @override_settings(AUTH_CONFIRM_EMAIL_ENABLE=True)
+    def test_basic_auth_unconfirmed(self):
+        url = '/api/v1.1/auth/login/'
+
+        resp = self.client.post(
+            url, {"email": 'juliadoe@netsach.org', "password": "plop"}
+        )
+        self.assertEqual(resp.status_code, status.HTTP_412_PRECONDITION_FAILED)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['EMAIL_NOT_VALIDATED'])
+
+    @override_settings(PASSWORD_EXPIRY_TIME=1)
+    def test_basic_auth_password_expired(self):
+        url = '/api/v1.1/auth/login/'
+
+        User = get_user_model()
+        user = User.objects.create(email='joho@netsach.org')
+        user.set_password('plop')
+        user.password_modification_date = pendulum.today().subtract(days=2)
+        user.save()
+        resp = self.client.post(
+            url, {"email": 'joho@netsach.org', "password": "plop"}
+        )
+        self.assertEqual(resp.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['EXPIRED_PASSWORD'])