+++ tests/tests_api_v1_1/test_api_v1_1_paginated_views.py
@@ -32,6 +32,8 @@ class TestPaginatedViews(APITestCase):
             get_url, HTTP_AUTHORIZATION='Token {}'.format(self.token)
         )
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['INVALID_QUERY'])
 
         pagination = 0
         get_url = '/api/v1.1/project/?c_resp_page_size={}'.format(pagination)
@@ -39,6 +41,8 @@ class TestPaginatedViews(APITestCase):
             get_url, HTTP_AUTHORIZATION='Token {}'.format(self.token)
         )
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['INVALID_QUERY'])
 
         pagination = 5
         get_url = '/api/v1.1/project/?c_resp_page_size={}'.format(pagination)
@@ -71,6 +75,8 @@ class TestPaginatedViews(APITestCase):
             get_url, HTTP_AUTHORIZATION='Token {}'.format(self.token)
         )
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['INVALID_QUERY'])
 
         nested = 'false'
         get_url = '/api/v1.1/project/?c_resp_nested={}'.format(nested)